Move elastic code to temba.contacts.search.elastic
