+++ temba/contacts/models.py
@@ -31,14 +31,14 @@ from temba.channels.models import Channel, ChannelEvent
 from temba.locations.models import AdminBoundary
 from temba.mailroom import ContactSpec, modifiers, queue_populate_dynamic_group
 from temba.orgs.models import Org, OrgLock
-from temba.utils import chunk_list, es, format_number, on_transaction_commit
+from temba.utils import chunk_list, format_number, on_transaction_commit
 from temba.utils.export import BaseExportAssetStore, BaseExportTask, TableExporter
 from temba.utils.models import JSONField as TembaJSONField, RequireUpdateFieldsMixin, SquashableModel, TembaModel
 from temba.utils.text import truncate, unsnakify
 from temba.utils.urns import ParsedURN, parse_urn
 from temba.values.constants import Value
 
-from .search import SearchException, parse_query
+from .search import SearchException, elastic, parse_query
 
 logger = logging.getLogger(__name__)
 
@@ -824,28 +824,6 @@ class Contact(RequireUpdateFieldsMixin, TembaModel):
             obj["urns"] = urns
 
         return obj
-
-    @classmethod
-    def query_elasticsearch_for_ids(cls, org, query, group=None):
-        from temba.contacts import search
-
-        try:
-            parsed = search.parse_query(org, query, group=group)
-            results = (
-                es.ModelESSearch(model=Contact, index="contacts")
-                .source(include=["id"])
-                .params(routing=org.id)
-                .using(es.ES)
-                .query(parsed.elastic_query)
-            )
-            matches = []
-            for r in results.scan():
-                matches.append(int(r.id))
-
-            return matches
-        except search.SearchException:
-            logger.error("Error evaluating query", exc_info=True)
-            raise  # reraise the exception
 
     def get_scheduled_messages(self):
         from temba.msgs.models import SystemLabel
@@ -2085,7 +2063,7 @@ class ExportContactsTask(BaseExportTask):
         include_group_memberships = bool(self.group_memberships.exists())
 
         if self.search:
-            contact_ids = Contact.query_elasticsearch_for_ids(self.org, self.search, group)
+            contact_ids = elastic.query_contact_ids(self.org, self.search, group=group)
         else:
             contact_ids = group.contacts.order_by("name", "id").values_list("id", flat=True)
