+++ temba/contacts/search/__init__.py
@@ -1,1 +1,1 @@
-from .base import ParsedQuery, SearchException, SearchResults, parse_query, search_contacts  # noqa
+from .mailroom import ParsedQuery, SearchException, SearchResults, parse_query, search_contacts  # noqa
